fix: UE 5.7 compatibility fixes
- Wrap FHttpServerResponse::Create TEXT() args in FString() (API change)
- Capture LexToString result to local FString to avoid dereference ambiguity
- Cast NodePosX/Y to int32 (type changed from double to int32 in 5.7)
- Use auto& for property binding path segments (type renamed from
  FStateTreePropertyPathSegment to FPropertyBindingPathSegment in 5.7)
- Add forward declaration of FClaireonServer

All changes are backwards-compatible with UE 5.5.

+++ Claireon/Source/Claireon/Private/Tools/ClaireonTool_EditBlueprintGraph.cpp
@@ -3091,7 +3091,7 @@ FToolResult ClaireonTool_EditBlueprintGraph::BuildStateResponse(const FString& S
 					}
 				}
 
-				StatusText += FString::Printf(TEXT("Position: (%.0f, %.0f)\n"), FocusedNode->NodePosX, FocusedNode->NodePosY);
+				StatusText += FString::Printf(TEXT("Position: (%d, %d)\n"), static_cast<int32>(FocusedNode->NodePosX), static_cast<int32>(FocusedNode->NodePosY));
 			}
 			else
 			{
@@ -3119,11 +3119,11 @@ FToolResult ClaireonTool_EditBlueprintGraph::BuildStateResponse(const FString& S
 			FString NodeTitle = Node->GetNodeTitle(ENodeTitleType::ListView).ToString();
 			bool bIsCursor = (Node->NodeGuid == Data->Cursor.FocusedNodeGuid);
 
-			StatusText += FString::Printf(TEXT("%d. [%s] @ (%.0f, %.0f)%s\n"),
+			StatusText += FString::Printf(TEXT("%d. [%s] @ (%d, %d)%s\n"),
 				NodeIndex++,
 				*NodeTitle,
-				Node->NodePosX,
-				Node->NodePosY,
+				static_cast<int32>(Node->NodePosX),
+				static_cast<int32>(Node->NodePosY),
 				bIsCursor ? TEXT(" <<<CURSOR>>>") : TEXT(""));
 
 			// Show execution connections (simplified)
